+++ packages/faas-cli-plugin-dev-pack/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.88](https://github.com/midwayjs/midway-faas/compare/v0.2.87...v0.2.88) (2020-04-26)
+
+**Note:** Version bump only for package @midwayjs/fcli-plugin-dev-pack
+
+
+
+
+
+## [0.2.87](https://github.com/midwayjs/midway-faas/compare/v0.2.86...v0.2.87) (2020-04-26)
+
+**Note:** Version bump only for package @midwayjs/fcli-plugin-dev-pack
+
+
+
+
+
+## [0.2.86](https://github.com/midwayjs/midway-faas/compare/v0.2.85...v0.2.86) (2020-04-24)
+
+**Note:** Version bump only for package @midwayjs/fcli-plugin-dev-pack
+
+
+
+
+
+## [0.2.85](https://github.com/midwayjs/midway-faas/compare/v0.2.84...v0.2.85) (2020-04-23)
+
+**Note:** Version bump only for package @midwayjs/fcli-plugin-dev-pack
+
+
+
+
+
+## [0.2.84](https://github.com/midwayjs/midway-faas/compare/v0.2.83...v0.2.84) (2020-04-23)
+
+**Note:** Version bump only for package @midwayjs/fcli-plugin-dev-pack
+
+
+
+
+
 ## [0.2.83](https://github.com/midwayjs/midway-faas/compare/v0.2.82...v0.2.83) (2020-04-23)
 
 **Note:** Version bump only for package @midwayjs/fcli-plugin-dev-pack